Composer: Fix SNYK-PHP-LARAVELFRAMEWORK-9292518
Overview
laravel/framework is a PHP framework for web artisans.

Affected versions of this package are vulnerable to Improper Neutralization in Validator.php via wildcard validation for file or image fields, such as files.*. This allows a user to bypass validation rules.

Remediation
Upgrade laravel/framework to version 11.44.1, 12.1.1 or higher.

Signed-off-by: Dicky Herlambang (花） <herlambangdicky5@gmail.com>